+++ Eye Strain/geometry.h
@@ -5,10 +5,10 @@
 
 double ConvertColor(int rgbValue);
 void DrawRect(Vector2 position, int width, int height);
-void DrawRect(Vector2 position, int width, int height, double color[3]);
+void DrawRect(Vector2 position, int width, int height, int color[3]);
 void DrawLine(Vector2 a, Vector2 b);
-void DrawLine(Vector2 a, Vector2 b, double color[3]);
+void DrawLine(Vector2 a, Vector2 b, int color[3]);
 void DrawCircle(Vector2 position, float radius);
-void DrawCircle(Vector2 position, float radius, double color[3]);
+void DrawCircle(Vector2 position, float radius, int color[3]);
 void DrawTriangle(Vector2 position, int width, int height);
 void DrawTriangle(Vector2 position, int width, int height, double angle);
